src/polygenic_adaptation/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"


[project]
name = "polygenic-adaptation"
authors = [
  { name = "Adam G. Fine", email = "afine2@uchicago.edu" },
]
description = "A framework for detecting polygenic adaptation from time-series genetic data"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.11"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = ["numpy>=1.21.0,<2.0","scipy>=1.11","matplotlib","numba>=0.59","tqdm","colorcet","Cartopy>=0.23","jinja2","cycler","seaborn>=0.13","scikit-allel>=1.3","snakemake>=8.0","snakemake-executor-plugin-cluster-generic","joblib>=1.4.2"]

[project.optional-dependencies]
test = [
  "pytest >=6",
  "pytest-cov >=3",
]
dev = [
  "pytest >=6",
  "pytest-cov >=3",
]
docs = [
  "sphinx>=7.0",
  "myst_parser>=0.13",
  "sphinx_copybutton",
  "sphinx_autodoc_typehints",
  "furo>=2023.08.17",
]

[project.urls]
Homepage = "https://github.com/enifmada/polygenic-adaptation"
"Bug Tracker" = "https://github.com/enifmada/polygenic-adaptation/issues"
Discussions = "https://github.com/enifmada/polygenic-adaptation/discussions"
Changelog = "https://github.com/enifmada/polygenic-adaptation/releases"


[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/polygenic_adaptation/_version.py"

[tool.hatch.envs.default]
features = ["test"]
scripts.test = "pytest {args}"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "ignore:::.*.numba",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]
pythonpath = "."


[tool.coverage]
run.source = ["polygenic_adaptation"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]


[tool.ruff]
src = ["src"]
exclude=["extract_vcfs/gdc"]
line-length = 120

[tool.ruff.lint]
extend-select = [
  "B",        # flake8-bugbear
  "I",        # isort
  "ARG",      # flake8-unused-arguments
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "ICN",      # flake8-import-conventions
  "G",        # flake8-logging-format
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
  "EXE",      # flake8-executable
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
]
isort.required-imports = ["from __future__ import annotations"]
# Uncomment if using a _compat.typing backport
# typing-modules = ["polygenic_adaptation._compat.typing"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]
"extract_vcfs/**" = ["T20"]

[tool.pylint]
py-version = "3.12.2"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "missing-function-docstring",
  "wrong-import-position",
]
